requirements.txt
mediapipe==0.10.20
numpy==2.2.1
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.0.0
ultralytics==8.3.55
